Ressourcen Controller, Mapper, Repository, Service & DTO erstellt. PersonController und Service geupdated

+++ backend/src/main/java/at/kaindorf/backend/model/Ressource.java
@@ -23,5 +23,5 @@ public class Ressource {
 
     private boolean isAvailable;
 
-    private int inventarNr;
+    private long inventarNr;
 }
